refactor: move negotiation into it's own module
